Add support for bLicenseMayNeedToBeUpdated
* Warns user when they have a license that appers to be outdated

This covers situations where the number of products covered by the license has been changed on the server because a new version of the software uses different/more products. In such cases, the user needs to download a new license file that includes these products in order to be able to use the new version.

* In addition, if there is a valid license, warnings fot that license are now reported as I intended; the code failed to report these before.